pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twinizer"
version = "0.1.47"
description = "A comprehensive environment for hardware and firmware reverse engineering, analysis and management"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Embedded Systems",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "prompt-toolkit>=3.0.0",
    "click>=8.1.8",
    "rich>=12.5.1,<13.0.0",
    "pyelftools>=0.27",
    "capstone>=4.0.2",
    "pefile>=2022.5.30",
    "decompyle3>=3.9.0",
    "pybind11>=2.10.0",
    "pdfalyzer>=0.9.0",
    "pdf2image>=1.16.0",
    "pdfminer.six>=20221105",
    "pypdf>=3.5.0",
    "python-magic>=0.4.27",
    "opencv-python>=4.5.0",
    "pytesseract>=0.3.10",
    "pillow>=9.0.0",
    "jinja2>=3.0.0",
    "platformio>=6.1.6",
    "pymupdf>=1.20.2",
    "ctypesgen>=1.1.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.61.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/twinizer"
"Bug Tracker" = "https://github.com/yourusername/twinizer/issues"

[project.scripts]
twinizer = "twinizer.__main__:main"